requirements.txt
alembic==1.0.0
Cython==0.28.5
git+https://github.com/cockroachdb/cockroachdb-python.git@298b70ce98459fd3ce43aa843ec2b3b16098df7b
asn1crypto==0.24.0
cryptography==2.3.1
falcon==1.4.1
gunicorn==19.9.0
jsonschema==2.6.0
passlib==1.7.1
PyJWT==1.6.4
python-jose==3.0.1
PyYAML==3.12
psycopg2==2.7.5
retrying==1.3.3
setproctitle==1.1.10
sqlalchemy==1.2.12
sqlalchemy-utils==0.33.5
